Regulation: preparation for nested PID
Flightcontrol regulates now the rate (angular speed) insteat of position.
This will allow the rate flight mode and better regulation for position.

Output to motors is filter with a median (kernel: 3) filter to avoid peaks caused by derivation in regulation (PID)
It could also be implemented on all inputs that are derivated or the errors.